refactor SeafoodMealsUI and SeafoodMealsUITest for improved package structure and readability
